moved infosection folder inside components folder
