+++ kandr5starevents/src/pages/index.js
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import HeroSection from "../components/HeroSection";
 import Navbar from "../components/Navbar";
 import Services from "../components/Services";
-import InfoSection from "../InfoSection";
+import InfoSection from "../components/InfoSection";
 import Sidebar from "../components/Sidebar";
 import { homeObjOne, homeObjTwo, homeObjThree } from "../InfoSection/Data";
 import Footer from "../components/Footer";
